+++ app/components/indicators/LoadingMoreTransactions.jsx
@@ -1,4 +1,4 @@
-import StakeyBounceXs from "./StakeyBounceExtraSmall";
+import StakeyBounceXs from "./StakeyBounce/StakeyBounceExtraSmall";
 import { FormattedMessage as T } from "react-intl";
 import styles from "./indicators.module.css";
